Move column writers to Arrow.scala
