+++ sql/core/src/main/scala/org/apache/spark/sql/Arrow.scala
@@ -21,14 +21,16 @@ import scala.collection.JavaConverters._
 import scala.language.implicitConversions
 
 import io.netty.buffer.ArrowBuf
-import org.apache.arrow.memory.RootAllocator
+import org.apache.arrow.memory.{BaseAllocator, RootAllocator}
+import org.apache.arrow.vector.{BaseDataValueVector, BitVector, IntVector, VarBinaryVector}
+import org.apache.arrow.vector.BaseValueVector.BaseMutator
 import org.apache.arrow.vector.schema.{ArrowFieldNode, ArrowRecordBatch}
 import org.apache.arrow.vector.types.FloatingPointPrecision
 import org.apache.arrow.vector.types.pojo.{ArrowType, Field, Schema}
 
-import org.apache.spark.sql.arrow._
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.types._
+import org.apache.spark.unsafe.types.UTF8String
 
 object Arrow {
 
@@ -106,4 +108,74 @@ object Arrow {
     val nullable = sparkField.nullable
     new Field(name, nullable, sparkTypeToArrowType(dataType), List.empty[Field].asJava)
   }
+}
+
+trait ColumnWriter {
+  def init(initialSize: Int): Unit
+  def write(data: Any): Unit
+  def finish(): (Seq[ArrowFieldNode], Seq[ArrowBuf])
+}
+
+/**
+ * Base class for flat arrow column writer, i.e., column without children.
+ */
+abstract class PrimitiveColumnWriter(protected val allocator: BaseAllocator) extends ColumnWriter {
+  protected val validityVector = new BitVector("validity", allocator)
+  protected val validityMutator = validityVector.getMutator
+  protected val valueVector: BaseDataValueVector
+  protected val valueMutator: BaseMutator
+
+  var count = 0
+  var nullCount = 0
+
+  protected def writeNull()
+  protected def writeData(data: Any)
+  protected def valueBufs(): Seq[ArrowBuf] = List(valueVector.getBuffer)
+
+  override def init(initialSize: Int): Unit = {
+    validityVector.allocateNew(initialSize)
+    valueVector.allocateNew()
+  }
+
+  override def write(data: Any): Unit = {
+    if (data == null) {
+      validityMutator.setSafe(count, 0)
+      writeNull()
+      nullCount += 1
+    } else {
+      validityMutator.setSafe(count, 1)
+      writeData(data)
+    }
+
+    count += 1
+  }
+
+  override def finish(): (Seq[ArrowFieldNode], Seq[ArrowBuf]) = {
+    validityMutator.setValueCount(count)
+    valueMutator.setValueCount(count)
+
+    val fieldNode = new ArrowFieldNode(count, nullCount)
+    (List(fieldNode), validityVector.getBuffer +: valueBufs)
+  }
+}
+
+class IntegerColumnWriter(allocator: BaseAllocator) extends PrimitiveColumnWriter(allocator) {
+  override protected val valueVector: IntVector = new IntVector("IntValue", allocator)
+  override protected val valueMutator: IntVector#Mutator = valueVector.getMutator
+
+  override protected def writeNull(): Unit = valueMutator.set(count, 0)
+  override protected def writeData(data: Any): Unit
+  = valueMutator.setSafe(count, data.asInstanceOf[Int])
+}
+
+class UTF8StringColumnWriter(allocator: BaseAllocator) extends PrimitiveColumnWriter(allocator) {
+  override protected val valueVector: VarBinaryVector
+  = new VarBinaryVector("UTF8StringValue", allocator)
+  override protected val valueMutator: VarBinaryVector#Mutator = valueVector.getMutator
+
+  override protected def writeNull(): Unit = valueMutator.setSafe(count, Array.empty[Byte])
+  override protected def writeData(data: Any): Unit
+  = valueMutator.setSafe(count, data.asInstanceOf[UTF8String].getBytes)
+  override protected def valueBufs(): Seq[ArrowBuf]
+  = List(valueVector.getOffsetVector.getBuffer, valueVector.getBuffer)
 }
